uart.h You shouldn't have to do anything with this.  It is included in Bluetooth.cpp for use.
uart.cpp
You shouldn't have to worry about this either.  It's the code for the lowest level of UART transmission, but I abstracted it away for you.

Bluetooth.cpp
the C++ code for our implementation of actually setting up UART transmission.  What you'll need to do in here is look at lines 178-191 where the main() method is, and move it into our game's main file. I put a large comment from lines 167-176 so you can see what exactly is going on there.  After that, you can delete the main() (since our program can't have 2 mains) but leave the rest of the code intact, obviously.

Bluetooth.h
This is the headerfile you need to include wherever you want to call the functions that are shown in Bluetooth.cpp.  Refer to the file for comments, but there are only 3 you should have to use:
void initializeBluetoothCOMPort(void);
void closeBluetoothCOMPort(void);
void pollForBluetoothData(void);

BluetoothGameData.h
This file is commented for you also.  This is what you'll have to include wherever you want to do some sort of check about game logic (ammo, etc) or sending of data to the microcontroller.  It's where you'll be calling most of the methods from to do game checks.

BluetoothGameData.cpp
Shouldn't have to worry about this, just the code for the methods.