Split up HighScores classes to separate files (ScoreItem and HighScoreList).
Fixed technical issue for FindRed that made it so that values that are not red be picked up by the webcam. Also optimized the memory management within the code in this file.

All graphics are now drawn as outlines instead of whole figures.

High scores are now displayed on StartScreen. Needs to be changed so that it is placed correctly on screen.

Basic code structure is written to allow the user's score to be tracked when a game session starts.